% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{mcg_run_bg}
\alias{mcg_run_bg}
\title{Start shiny app in background}
\usage{
mcg_run_bg(data, port = 1151)
}
\arguments{
\item{data}{myClim object}

\item{port}{of app (default 1151)}
}
\value{
An r_process object see \code{\link[callr:r_bg]{callr::r_bg()}}
}
\description{
This function starts the Shiny application in new process. The R console doesn't freeze.
}
\details{
The function mcg_run_bg doesn't open automatically app. You must open it manually in browser
with link http://localhost:\if{html}{\out{<port number>}}. The result of function is r_process object see \code{\link[callr:r_bg]{callr::r_bg()}}.
You can get myClim object returned from app by call function \code{r_process$get_result()}.
}
\examples{
\dontrun{
     proc <- myClimGui::mcg_run_bg(myClim::mc_data_example_agg)
     # after click Return button
     data <- proc$get_result()
}
}
