% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myClimGui-package.R
\docType{package}
\name{myClimGui-package}
\alias{myClimGui}
\alias{myClimGui-package}
\title{myClimGui: GUI fo myClim package}
\description{
Shiny app for myClim
}
\author{
\strong{Maintainer}: Vojtěch Kalčík \email{Vojtech.Kalcik@ibot.cas.cz}

Authors:
\itemize{
  \item Matěj Man \email{Matej.Man@ibot.cas.cz}
  \item Martin Macek \email{Martin.Macek@ibot.cas.cz}
  \item Josef Brůna \email{Josef.Bruna@ibot.cas.cz}
  \item Anna Růžičková \email{Anna.Ruzickova@ibot.cas.cz}
  \item Jan Wild \email{Jan.Wild@ibot.cas.cz}
  \item Martin Kopecký \email{Martin.Kopecky@ibot.cas.cz}
}

Other contributors:
\itemize{
  \item Institute of Botany of the Czech Academy of Sciences [copyright holder]
}

}
